+++ app/(menu)/site/[name]/pageWrapperHeader.tsx
@@ -3,7 +3,7 @@
 import React from "react";
 import { useRouter } from "next/navigation";
 import { DetailFormOption } from "@/types/form";
-import RoundButton from "@/components/btn/RoundButton";
+import Button from "@/components/btn/Button";
 
 export default function PageWrapperHeader({
   option,
@@ -22,7 +22,7 @@ export default function PageWrapperHeader({
   return (
     <div className="w-full grid grid-cols-3 p-3 text-black sm:w-4/5 sm:px-1">
       <div className="flex justify-start items-end col-span-1">
-        <RoundButton onClick={() => backBtnOnClick(url)} text="<" />
+        <Button onClick={() => backBtnOnClick(url)} text="<" type="ROUNDED"/>
       </div>
       <div className="flex items-center justify-center col-span-1 text-center">
         {forName}
