+++ components/form/ListForm.tsx
@@ -16,7 +16,7 @@ export default function ListForm<T>({
   const list: List<T> = form.list;
 
   const contents: T[] = list.contents;
-  const classes = "w-full grid grid-cols-12 text-black gap-2 px-3 py-2 pb-0";
+  const classes = "w-full grid grid-cols-12 text-black gap-2  pb-0";
   return (
     <>
       {/* Header */}
